Refactoring code to get ready to merge to master
