+++ main/src/main/scala/org/clulab/reach/context/ContextEngineFactory.scala
@@ -1,7 +1,5 @@
 package org.clulab.reach.context
 
-import org.clulab.reach.context.context_exec.SVMContextEngine
-
 
 object ContextEngineFactory {
